Rename custom ESLint rule to prefer-use-i18n-polyfill
